requirements.txt
asgiref==3.7.2
cloudinary==1.33.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.19
django-allauth==0.54.0
django-filter==23.2
django-widget-tweaks==1.4.12
gunicorn==20.1.0
oauthlib==3.2.2
psycopg2==2.9.6
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.15
